requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pillow==9.5.0
pandas==2.0.2
ttkthemes~=3.2.2